♻️ Refactor: callchat 리포지터리 파일 이동
