+++ src/main/java/callprotector/spring/domain/chat/service/CallChatLogServiceImpl.java
@@ -4,7 +4,7 @@ import callprotector.spring.domain.callchat.dto.response.CallChatLogResponseDTO;
 import callprotector.spring.domain.chat.dto.response.SourcePageDTO;
 import callprotector.spring.domain.callchat.entity.CallChatLog;
 import callprotector.spring.domain.callchat.entity.CallChatSession;
-import callprotector.spring.domain.chat.repository.CallChatLogRepository;
+import callprotector.spring.domain.callchat.repository.CallChatLogRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
